+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 + Added `JImTexture.fromBytes` for in-memory image loading
 + Added constructor for `JImGui` which takes a `long` as the native object pointer
   + This requires you to manually manage the release and initialization of GLFW
++ Renamed some factories of `JImTexture`
++ Added `windowDrawListAddLine`, `windowDrawListAddImage`
 
 ## v0.4